Linux compatibility & bug fixes

+++ source/main.cc
@@ -63,7 +63,7 @@ int main (int argc, char** args) {
 
   
   I.run();
-  printf("fib_i(%d) = %llu\n", N, *(uint64_t*) I.op_registers + RAX);
+  printf("fib_i(%d) = %zu\n", N, *(size_t*) I.op_registers + RAX);
   printf("fib_s(%d) = %d\n", N, fib_s(N, (void*) fib_s));
